+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # 2024/12/05
 * Finished implementing the SFEN outputter 
+
+# 2024/12/06
+* Implemented the SFEN parser. Only remaining part is to
+  * add some integration tests of SFEN -> parse -> display -> SFEN